adaptation_ser/requirements.txt
 ruamel.yaml
 librosa
 numpy
 tensorflow